Allow live-reload
This updates can-ssr to allow for live-reload. To make this work we are
now polyfilling the global.WebSocket constructor and providing a way to
configure the hostname used to connect to the live-reload server.

Import: this change is going to require Steal 0.12.0 to work. This is
because it includes changes to live-reload (and system-trace) that are
necessary for server side live-reload to work.

Fixes #14