Updated live-reload tests
This refactors how the live-reload tests work making them a bit better.
The way it works is we start a script that launches a `live-reload-test`
server and a `can-serve` server on a project. The test uses Steal to
load `live-reload-testing` with the WebSocket polyfill so the module can
be used the same way as in the client. Then we just modify the project
and curl the result to see if it changed correctly.