Implements new Main Module API
This commit implements the Main Module API as defined in #80. What this
means is that you can use can-ssr with any JavaScript module as long as
they implement these two APIs:

Creates an Object that will hold state relevant for the request. This
could be some type of typed object that acts as a view model for your
app or it could be a plain object, depending on what your framework
needs.

```js
exports.createState = function(){
	return new AppViewModel();
};
```

**render** is a function that returns a Promise when it has finished
rendering. It is given a new Document that it can write to, or it could
also return a string representation.

```js
exports.render = function(doc, state){
	var $body = $(doc.body);

	$body.append($("<div>Hello world</div>"));

	return $.getJSON("/some/data.json").then(function(data){
		// render the data some how
	});
};
```

If writing directly to the provided Document can-ssr will stringify the
documentElement and return that as the response.