Override can.document when rendering
Previously done-autorender handled setting can.document to the request's
document by waiting to insert the fragment until it had been fully
rendered. This meant that can.document was only changed once and
inserted events were only fired once.

This was a bug because it meant if async activity was triggered by
something that happened within an `inserted` event, that activity would
not be rendered on the server.

With the change to use can-wait we need to be able to set can.document
every time a different task is performed (a setTimeout, a Promise,
		whatever). can-wait supports passing in an array of Overrides
that will be set for each task. So the fix is to pass in an override
that overrides can.document.