Return a Stream instead of a Promise
Closes #145

This changes the core API to return a stream instead of a Promise.
Coupel of reasons this is needed:

1) Previously we didn't have the Response object, which is needed for
adding cookies that happened as a result of a request. Changing to
streams means the response is piped to, so we can write to it. This is
also true of the statusCode.

2) Eventually we want to make rendering be streaming, so this change
would be needed then anyways.