Remove server and middleware from the project
This removes the server (and can-serve) and middleware parts from the
project, they have been moved to done-serve and done-ssr-middleware
respectively.

Also gets rid of Grunt and replaces with a script which does the copying
for tests. Cleans up the mocha running so it's easier to add new tests
(just edit `test/test.js`).

Explains a little bit more what is needed for the main module but still
links to the docs specifically for that.

Adds a note explaining where the other projects have moved to.

Closes #127