Allow configuring an authCookie name for token auth
This adds an option to the xhr zone that allows you to add a cookie name where an auth token will be potentially located.  If a token is found, it copies the token to the authorization header of all outgoing requests.

Also removed the token_test and set up the auth-cookie_test, since this replaces that functionality.