Use can.route.matched() to determine if a route matched
This changes the behavior of 404s, so that we use the
can.route.matched() compute to determine if a route was matched when
rendering. This is used to determine which `statusCode` is used, if the
user doesn't provide it themselves on their ViewModel.

Also adds a few tests to make sure this works with DefineMaps. Closes