Allow main to be executed as the "render" function
When the main module doesn't provide an exported function, assume that
re-executing the main module's body is how rendering should occur.

Note that this currently only works for ES modules. This is because our
POC uses ES modules and I want to prove out the concept first before
proceeding with dynamic module types. Will create an issue to follow
back and get those working in this situation as well.

Closes #368