Make sure steal-css appends to the global document
A big change in the refactoring of done-ssr into zones is that now steal
loads during the first request; it doesn't load *prior* to the first
request. That means that side-effects of loading steal, in the case of
steal-css is that it appends styles/links to the document, are going to
be part of the user's rendered result.

This is undesirable. Lucky steal-css checks for a
`doneSsr.globalDocument` and will append to that instead. This change
makes it so that this document is created and used.

Closes #386