Use can-globals/document/ and URL constructor
This is two fixes that fix the tests:

1. Using can-globals/document/document instead of
can-util/dom/document/document. For some reason the wrong result of
getDocument() was returning when using the latter. Rather than
investigate, just don't going to use the deprecated module any more.
2. Use the URL constructor to make the `location` object, since that
will be the same as the real `location` object. For example,
	 `location.search` will be an empty string when there is no search,
	 rather than null like was working before.