Ensure the reattachment script's parent is the <head>
This ensures that the reattachment script (contained with the iframe) is
within the `<head>`. This is necessary so that the server DOM and the
iframe's DOM match. Otherwise the mutation commands would be off.

This change is necessary because most index.stache files contain a
TextNode beneath the document element.

Fixes #485