Support fetch() response.ok
The fetch API has a `response.ok` property that signifies a 20x
response.

This is supported by node-fetch. However we were making a copy of the
`response` object, and since `ok` is implemented as a non-enumerable
getter, our response object didn't contain it. The fix is to instead use
`Object.create()` to create a new object that extends from node-fetch's
response, and therefore inheriting it's prototype chain.

Fixes #509