Prevent leak caused by saving promise with can-dom-data-state
This prevents a leak caused by storing a promise with
can-dom-data-state. Since this instance of can-dom-data-state comes from
Node's module system and not Steal's, this doesn't get cleaned up when
we clean up memory with can-dom-mutate.

In this case we can replace the usage with Symbols.

This commit also turns back on the memory leak test, which is now
passing. Closes #560