Leave on a documentElement on the document.
We remove the `documentElement` in order to allow all of memory cleanup
to occur. However some libraries, like jQuery freak out when a document
doesn't have a documentElement.

The solution is to instead replace the documentElement with a new one.
This means there still is one, but the old one can have its memory
cleaned up.